+++ packages/index/src/write.js
@@ -1,4 +1,4 @@
-import fs from 'fs/promises';
+import path from 'path';
 
 import { writeJsonGzip, promisify } from '@home-gallery/common';
 
@@ -6,15 +6,15 @@ import { byDirDescFileAsc } from './utils.js';
 
 const asyncWriteJsonGzip = promisify(writeJsonGzip)
 
-export const writeIndex = (filename, index) => {
-  index.created = new Date().toISOString();
-  index.data.sort(byDirDescFileAsc);
-  const tmp = `${filename}.tmp`;
-  return asyncWriteJsonGzip(tmp, index)
-    .then(() => {
-      return fs.rename(tmp, filename)
-    })
-    .then(() => {
-      return index
-    })
+export const writeIndex = async (directory, filename, entries, options) => {
+  const index = {
+    type: 'home-gallery/fileindex@1.0',
+    created: new Date().toISOString(),
+    base: path.resolve(directory),
+    data: entries.sort(byDirDescFileAsc)
+  }
+  if (options.dryRun) {
+    return index
+  }
+  return asyncWriteJsonGzip(filename, index).then(() => index)
 }